feat: remove unused files, improve article seo, ready to add tags in article page
