+++ components/article/content/ArticleContentTags.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts"></script>
+
+<template>
+  <div>
+    I'm tags component
+  </div>
+</template>
+
+<style scoped></style>
