+++ package.json
@@ -2,7 +2,7 @@
   "name": "nuxt3-starter",
   "type": "module",
   "private": true,
-  "packageManager": "pnpm@10.8.1",
+  "packageManager": "pnpm@10.11.1",
   "scripts": {
     "preinstall": "npx only-allow pnpm",
     "build": "nuxt build",
