+++ utils/date.ts
@@ -13,5 +13,5 @@ export function ISODate(date: string | Date): string {
 
   const dayjs = useDayjs()
 
-  return dayjs.tz(date).toISOString()
+  return dayjs.tz(date).format()
 }
